refactor(spec): split temporal-parse into core, format, and named engines
Extract calendar foundation and the exact-format compile cache from the
value-level parser so engines can evolve independently. Co-locate exact-format
and tier-2 temporal scale tests. Public parseTemporal behavior is unchanged.
