+++ .changeset/spec-temporal-parse-engines-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split value-level temporal parsing into core, exact-format, and named-engine modules, and co-locate format plus tier-2 temporal scale tests. Public parseTemporal surface is unchanged.
